FIX:add time and card existence checking logic in card.service.js

+++ src/routes/card/card.service.js
@@ -27,6 +27,11 @@ export class CardsService {
     cardData.cardStartTime = getDateTimeFormat(cardData.cardStartTime);
     //종료시간의 시간 형식을 변경
     cardData.cardEndTime = getDateTimeFormat(cardData.cardEndTime);
+    if (cardData.cardStartTime > cardData.cardEndTime) {
+      const error = new Error('시작 시간은 종료시간보다 빠를 수 없습니다.');
+      error.status = 400;
+      throw error;
+    }
     const lastCardOrder = await this.CardsRepository.findLastCardOrder(
       cardData.columnId
     );
@@ -37,22 +42,46 @@ export class CardsService {
     return card;
   };
   updateCard = async (cardId, cardWriterId, cardData) => {
+    const targetCard = await this.CardsRepository.findCard(cardId);
+    if (!targetCard) {
+      const error = new Error('카드가 존재하지 않습니다.');
+      error.status = 404;
+      throw error;
+    }
+    const cardStartTime = targetCard.cardStartTime;
+    const cardEndTime = targetCard.cardEndTime;
     //시작시간을 수정하는 경우 시간 형식을 변경
     if (cardData.cardStartTime) {
-      cardData.cardStartTime = getDateTimeFormat(cardData.cardStartTime);
+      cardStartTime = getDateTimeFormat(cardData.cardStartTime);
     }
     //종료시간을 수정하는 경우 시간 형식을 변경
     if (cardData.cardEndTime) {
-      cardData.cardEndTime = getDateTimeFormat(cardData.cardEndTime);
+      cardEndTime = getDateTimeFormat(cardData.cardEndTime);
+    }
+    if (cardStartTime > cardEndTime) {
+      const error = new Error('시작 시간은 종료시간보다 빠를 수 없습니다.');
+      error.status = 400;
+      throw error;
     }
     const card = await this.CardsRepository.updateCard(
       cardId,
       cardWriterId,
       cardData
     );
+    if (card) {
+      const error = new Error('카드가 존재하지 않습니다.');
+      error.status = 404;
+      throw error;
+    }
     return card;
   };
   deleteCard = async (cardId) => {
+    const targetCard = await this.CardsRepository.findCard(cardId);
+    if (!targetCard) {
+      const error = new Error('카드가 존재하지 않습니다.');
+      error.status = 404;
+      throw error;
+    }
     const card = await this.CardsRepository.deleteCard(cardId);
     return card;
   };
